Add clear task to remove all completed tasks
