+++ src/index.js
@@ -1,9 +1,9 @@
 import './style.css';
 import RecycleImg from './recycle.svg';
 import MoreImg from './more.svg';
-import TrashImg from './delete.svg';
 import { drag, drop, allowDrop } from './drag_drop.js';
 import updateTasks from './status.js';
+import { addTask, editTask, clear } from './add_remove.js';
 
 let tasks = null;
 
@@ -19,76 +19,20 @@ window.updateLocalStorage = function updateLocalStorage(retrieve) {
   window.displayTasks();
 };
 
+/**       Handler for calling a task from and inline declared listener */
+window.callAddTask = function callAddTask() {
+  addTask(tasks);
+};
+
 /**       Update the state of the tasks            */
-window.update = function update() {
-  const response = updateTasks();
-  tasks = response;
-  window.updateLocalStorage(false);
-};
-
-/**       AddTask adds tasks to the tasks list      */
-window.addTask = function addTask() {
-  const str = document.getElementById('description').value;
-  const firstLetter = str.charAt(0).toUpperCase();
-  str.replace(str.charAt(0), firstLetter);
-  const description = str;
-  const completed = false;
-  const date = new Date();
-  const id = date.getMilliseconds();
-
-  if (!tasks) {
-    tasks = [];
+window.update = function update(data) {
+  if (!data) {
+    const response = updateTasks();
+    tasks = response;
+  } else {
+    tasks = data;
   }
 
-  const index = tasks.length + 1;
-
-  if (tasks && description !== '') {
-    const task = {
-      description,
-      completed,
-      index,
-      id,
-    };
-    tasks.push(task);
-    tasks.sort((taskA, taskB) => {
-      const indexA = taskA.position;
-      const indexB = taskB.position;
-      if (indexA < indexB) {
-        return -1;
-      }
-      if (indexA > indexB) {
-        return 1;
-      }
-      return 0;
-    });
-
-    window.updateLocalStorage(false);
-  }
-};
-
-window.editTask = function editTask(divId) {
-  const list = document.getElementsByClassName('drag-div');
-  Array.from(list).forEach((li) => {
-    if (li.id === divId) {
-      li.style.backgroundColor = '#fff59c78';
-      const img = li.getElementsByTagName('img')[0];
-      img.src = TrashImg;
-    } else {
-      li.style.backgroundColor = 'white';
-      const img = li.getElementsByTagName('img')[0];
-      img.src = MoreImg;
-    }
-  });
-};
-
-window.clear = function clear() {
-  const temp = [];
-  tasks.forEach((task) => {
-    if (task.completed !== true) {
-      temp.push(task);
-    }
-  });
-  tasks = temp;
   window.updateLocalStorage(false);
 };
 
@@ -102,7 +46,6 @@ window.displayTasks = function displayTasks() {
   if (tasks) {
     tasks.forEach((task, index) => {
       const { description, id } = task;
-      // const taskCard = `<li id=${liId} onclick="window.editTask(${liId})" draggable="true" >
       const li = document.createElement('li');
       li.id = index;
       li.addEventListener('drop', (EventTarget) => {
@@ -121,7 +64,7 @@ window.displayTasks = function displayTasks() {
       div.id = divId;
       div.classList.add('drag-div');
       div.draggable = true;
-      div.addEventListener('click', () => window.editTask(divId));
+      div.addEventListener('click', () => editTask(divId, tasks));
       div.data = index;
       div.addEventListener('dragstart', (EventTarget) => {
         div.classList.add('dragging');
@@ -166,7 +109,6 @@ window.displayTasks = function displayTasks() {
       list.appendChild(li);
     });
   }
-
   const template = `
   <div class="top">
   <h1 class="title">Today's To Do</h1>
@@ -175,7 +117,7 @@ window.displayTasks = function displayTasks() {
             <img class="add-btn-img" src=${RecycleImg} alt="" /> 
             </button>
   </div>       
-          <form onsubmit="window.addTask()" id="task-form">
+          <form onsubmit="window.callAddTask()" id="task-form">
             <input
               id="description"
               type="text"
@@ -192,8 +134,9 @@ window.displayTasks = function displayTasks() {
   container.innerHTML = template;
   const buttonHtml = document.createElement('button');
   buttonHtml.id = 'clear-btn';
-  buttonHtml.classList.add('clear-btn');
-  buttonHtml.onclick = 'window.clear()';
+  buttonHtml.addEventListener('click', () => {
+    clear(tasks);
+  });
   buttonHtml.textContent = 'Clear completed tasks.';
   container.insertAdjacentElement('beforeend', list);
   container.insertAdjacentElement('beforeend', buttonHtml);
